requirements/base.txt
# Django & DRF
Django==3.2.20
djangorestframework==3.14.0
django-cors-headers==4.1.0
django-filter==23.2
django-environ==0.10.0
markdown==3.4.3
drf-nested-routers==0.93.4
drf-yasg==1.21.7
django-rest-swagger==2.2.0

# Database
psycopg2-binary==2.9.6

# Channels
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Celery
celery==5.3.1
django-celery-beat==2.5.0
redis==4.6.0

# Storage & Media
Pillow==10.0.0

# Deployment
gunicorn==21.2.0
whitenoise==6.5.0